separate mirror-list
